gp_upgrade: Rename command_listener to agent
binary will be gp_upgrade_agent; all in-code references simply "agent"
